+++ NET1814_MilkShop.Repositories/Models/CheckoutModels/CheckoutModel.cs
@@ -4,7 +4,7 @@ namespace NET1814_MilkShop.Repositories.Models.CheckoutModels;
 
 public class CheckoutModel
 {
-    public decimal ShippingFee { get; set; }
+    public int ShippingFee { get; set; }
 
     [Required]
     public int AddressId { get; set; }
