+++ NET1814_MilkShop.Repositories/Models/ProductModels/ProductModel.cs
@@ -10,9 +10,9 @@
 
         public int Quantity { get; set; }
 
-        public decimal OriginalPrice { get; set; }
+        public int OriginalPrice { get; set; }
 
-        public decimal SalePrice { get; set; }
+        public int SalePrice { get; set; }
 
         public string? Thumbnail { get; set; }
