+++ NET1814_MilkShop.Repositories/Models/ProductModels/ProductQueryModel.cs
@@ -10,10 +10,10 @@ namespace NET1814_MilkShop.Repositories.Models.ProductModels
         public string? Status { get; set; }
 
         [Range(0, double.MaxValue, ErrorMessage = "Min price must be greater than 0")]
-        public decimal MinPrice { get; set; }
+        public int MinPrice { get; set; }
 
         [Range(0, double.MaxValue, ErrorMessage = "Max price must be greater than 0")]
-        public decimal MaxPrice { get; set; }
+        public int MaxPrice { get; set; }
         public bool? IsActive { get; set; }
 
         /// <summary>
